requirements.txt
setuptools
click
gitpython
pyyaml
kubernetes
dataclasses_json
